% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp2prot.R
\name{bp2prot}
\alias{bp2prot}
\title{Biological process to protein (bp2prot)}
\usage{
bp2prot(biological_process, taxon = NULL, sources = FALSE)
}
\arguments{
\item{biological_process}{Biological process label or ID (string). Example: "coenzyme A transmembrane transport".}

\item{taxon}{Taxon name (string) or taxon ID (string or integer). Example: "Homo sapiens" or 9606.}

\item{sources}{If sources want to be included (boolean: True or False).}
}
\value{
A data.frame of proteins associated with the biological process and their sources.
}
\description{
Function to obtain the proteins associated with a biological process.
}
\examples{
\dontrun{
bp2prot("GO:0043524", "Homo sapiens", sources = TRUE)
}

}
