% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc2prot.R
\name{cc2prot}
\alias{cc2prot}
\title{Cellular component to protein (cc2prot)}
\usage{
cc2prot(cellular_component, taxon = NULL, sources = FALSE)
}
\arguments{
\item{cellular_component}{Cellular component label or ID (string). Example: "cation channel complex".}

\item{taxon}{Taxon name (string) or taxon ID (string or integer). Example: "Homo sapiens" or 9606.}

\item{sources}{If sources want to be included (boolean: TRUE or FALSE).}
}
\value{
A data.frame of proteins associated with the cellular component and their sources if sources=TRUE, or a vector of protein names if sources=FALSE.
}
\description{
Function to obtain the proteins associated with a cellular component.
}
\examples{
\dontrun{
cc2prot("GO:0034703", "Homo sapiens", sources = TRUE)
}

}
