% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm2gene.R
\name{crm2gene}
\alias{crm2gene}
\title{CRM to gene (crm2gene)}
\usage{
crm2gene(crm)
}
\arguments{
\item{crm}{CRM label (string). Example: "crm/CRMHS00000005764".}
}
\value{
A data.frame of genes associated with the CRM and their sources.
}
\description{
Function to obtain the genes associated with a CRM (Cis-Regulatory Module).
}
\examples{
\dontrun{
crm2gene("crm/CRMHS00000005764")
}

}
