% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm2phen.R
\name{crm2phen}
\alias{crm2phen}
\title{CRM to phenotype (crm2phen)}
\usage{
crm2phen(crm)
}
\arguments{
\item{crm}{CRM label (string). Example: "crm/CRMHS00000007832".}
}
\value{
A data.frame of phenotypes associated with the CRM and their sources, ordered by phenotype ID.
}
\description{
Function to obtain the phenotypes associated with a CRM (Cis-Regulatory Module).
}
\examples{
\dontrun{
crm2phen("crm/CRMHS00000005764")
}

}
