% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm2tfac.R
\name{crm2tfac}
\alias{crm2tfac}
\title{CRM to transcription factor (crm2tfac)}
\usage{
crm2tfac(crm)
}
\arguments{
\item{crm}{CRM label (string). Example: "crm/CRMHS00000007832".}
}
\value{
A data.frame of transcription factors associated with the CRM and their sources, ordered by transcription factor name.
}
\description{
Function to obtain the transcription factors associated with a CRM (Cis-Regulatory Module).
}
\examples{
\dontrun{
crm2tfac("crm/CRMHS00000007832")
}

}
