% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2crm.R
\name{gene2crm}
\alias{gene2crm}
\title{Gene to CRM (gene2crm)}
\usage{
gene2crm(gene)
}
\arguments{
\item{gene}{Gene label (string). Example: "TOX3".}
}
\value{
A data.frame of CRMs associated with the gene and their sources, ordered by CRM name.
}
\description{
Function to obtain the CRMs (Cis-Regulatory Modules) associated with a gene.
}
\examples{
\dontrun{
gene2crm("TOX3")
}

}
