% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2phen.R
\name{gene2phen}
\alias{gene2phen}
\title{Gene to phenotype (gene2phen)}
\usage{
gene2phen(gene)
}
\arguments{
\item{gene}{Gene name (string). Example: "BRCA1".}
}
\value{
A data.frame with OMIM IDs and phenotype labels associated with the gene, ordered by phenotype label.
}
\description{
Function to retrieve phenotypes associated with a given gene.
}
\examples{
\dontrun{
gene2phen("BRCA1")
}

}
