% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2prot.R
\name{gene2prot}
\alias{gene2prot}
\title{Gene to proteins (gene2prot)}
\usage{
gene2prot(gene, taxon = NULL)
}
\arguments{
\item{gene}{Gene name (string). Example: "BRCA1".}

\item{taxon}{Taxon name (string) or taxon ID (string or integer), or NULL if not specified. Example: "Homo sapiens" or 9606. Default: NULL.}
}
\value{
A vector of protein names associated with the gene.
}
\description{
Function to retrieve proteins associated with a given gene, optionally filtered by taxon.
}
\examples{
\dontrun{
gene2prot("TOX3", taxon = "Homo sapiens")
}

}
