% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCRM_info.R
\name{getCRM_info}
\alias{getCRM_info}
\title{Get CRM information (getCRM_info)}
\usage{
getCRM_info(crm)
}
\arguments{
\item{crm}{Name of the CRM (string). Example: "crm/CRMHS00000005387".}
}
\value{
A data.frame containing the start, end, chromosome, assembly, taxon, and definition of the CRM. If no data is available, a message is returned.
}
\description{
Function to obtain the main data for a specific cis-regulatory module (CRM), including coordinates, taxon, and definition.
}
\examples{
\dontrun{
getCRM_info("crm/CRMHS00000005387")
}

}
