% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCRMs_by_overlap.R
\name{getCRMs_by_overlap}
\alias{getCRMs_by_overlap}
\title{Get CRMs by overlap (getCRMs_by_overlap)}
\usage{
getCRMs_by_overlap(chromosome, start, end = NULL)
}
\arguments{
\item{chromosome}{chromosome name (string). Example: "chr-1".}

\item{start}{start position (integer).}

\item{end}{end position (integer). Default: NULL}
}
\value{
A data.frame of overlapping CRMs.
}
\description{
Function to obtain the human cis-regulatory modules (CRMs) that completely overlap with a specific position or a specific range of genomic coordinates.
}
\examples{
\dontrun{
getCRMs_by_overlap("chr-1", 52565276)
}

}
