% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenes_by_coord.R
\name{getGenes_by_coord}
\alias{getGenes_by_coord}
\title{Get genes by coordinates (getGenes_by_coord)}
\usage{
getGenes_by_coord(chr, start, end, strand = NULL)
}
\arguments{
\item{chr}{Chromosome name (string). Example: "chr-1".}

\item{start}{Start position (integer).}

\item{end}{End position (integer).}

\item{strand}{Strand orientation (string), or NULL if not specified. Possible values: "ForwardStrandPosition", "ReverseStrandPosition" and NULL. Default: NULL.}
}
\value{
A data.frame containing gene names, start, end, and strand for genes within the specified coordinates. If no data is available, a message is returned.
}
\description{
Function to retrieve human genes located within a specified range of genomic coordinates, optionally filtered by strand orientation.
}
\examples{
\dontrun{
getGenes_by_coord("chr-1", 52565276, 58596412, "ForwardStrandPosition")
}

}
