% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenes_by_overlap.R
\name{getGenes_by_overlap}
\alias{getGenes_by_overlap}
\title{Get genes by ovelap (getGenes_by_overlap)}
\usage{
getGenes_by_overlap(chr, start, end = NULL, strand = NULL)
}
\arguments{
\item{chr}{Chromosome name (string). Example: "chr-1".}

\item{start}{Start position (integer).}

\item{end}{End position (integer). Default: NULL}

\item{strand}{Strand orientation (string), or NULL if not specified. Possible values: "ForwardStrandPosition", "ReverseStrandPosition" and NULL. Default: NULL.}
}
\value{
A data.frame containing gene names, start, end, and strand for overlapping genes. If no data is available, a message is returned.
}
\description{
Function to retrieve human genes that completely overlap with a specific position or a specific range of genomic coordinates, with the option to filter them by strand orientation.
}
\examples{
\dontrun{
getGenes_by_overlap("chr-16", 52565276)
}

}
