% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhenotype.R
\name{getPhenotype}
\alias{getPhenotype}
\title{Get phenotype information (getPhenotype)}
\usage{
getPhenotype(phenotype)
}
\arguments{
\item{phenotype}{Phenotype label (string), OMIM ID (6-digit string), or MTHU identifier (string). Example: "breast cancer" or "604704".}
}
\value{
A data.frame containing OMIM IDs and phenotype labels. If no data is available, a message is returned.
}
\description{
Function to retrieve phenotype information based on a given phenotype label, OMIM ID, or MTHU identifier.
}
\examples{
\dontrun{
getPhenotype("breast cancer")
}

}
