% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProtein_info.R
\name{getProtein_info}
\alias{getProtein_info}
\title{Get protein information (getProtein_info)}
\usage{
getProtein_info(protein, sources = T)
}
\arguments{
\item{protein}{Name of the protein (string). Example: "TBX5_HUMAN".}

\item{sources}{Include sources (boolean: TRUE or FALSE). Default: TRUE.}
}
\value{
A list containing unique protein alternative IDs, definition, evidence level, articles, and alternative sources. If no data is available, a message is returned.
}
\description{
Function to obtain the main features of a specific protein, including ID, definition, evidence level, and optionally articles and alternative sources.
}
\examples{
\dontrun{
getProtein_info("TBX5_HUMAN", sources = TRUE)
}

}
