% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTAD_add_info.R
\name{getTAD_add_info}
\alias{getTAD_add_info}
\title{Get additional TAD information (getTAD_add_info)}
\usage{
getTAD_add_info(tad)
}
\arguments{
\item{tad}{Name of the TAD (string). Example: "tad/TADHS00000038004".}
}
\value{
A list containing unique values for evidence, databases, biological samples, and articles associated with the TAD. If no data is available, a message is returned.
}
\description{
Function to obtain additional data for a specific topologically associating domain (TAD), including evidence, databases, biological samples, and articles.
}
\examples{
\dontrun{
getTAD_add_info("tad/TADHS00000038004")
}

}
