% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTAD_info.R
\name{getTAD_info}
\alias{getTAD_info}
\title{Get TAD information (getTAD_info)}
\usage{
getTAD_info(tad)
}
\arguments{
\item{tad}{Name of the TAD (string). Example: "tad/TADHS00000038004".}
}
\value{
A data.frame containing the chromosome, start, end, assembly, taxon, and definition of the TAD. If no data is available, a message is returned.
}
\description{
Function to obtain the main data for a specific topologically associating domain (TAD), including coordinates, taxon, and definition.
}
\examples{
\dontrun{
getTAD_info("tad/TADHS00000038004")
}

}
