% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTADs_by_coord.R
\name{getTADs_by_coord}
\alias{getTADs_by_coord}
\title{Get TADs by coordinates (getTADs_by_coord)}
\usage{
getTADs_by_coord(chromosome, start, end)
}
\arguments{
\item{chromosome}{Chromosome name (string). Example: "chr-13".}

\item{start}{Start position (integer).}

\item{end}{End position (integer).}
}
\value{
A data.frame containing TAD IDs, start, and end positions for TADs within the specified coordinates. If no data is available, a message is returned.
}
\description{
Function to retrieve human topologically associating domains (TADs) located within a specified range of genomic coordinates.
}
\examples{
\dontrun{
getTADs_by_coord("chr-13", 34120000, 35840000)
}

}
