% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTADs_by_overlap.R
\name{getTADs_by_overlap}
\alias{getTADs_by_overlap}
\title{Get TADs by overlap (getTADs_by_overlap)}
\usage{
getTADs_by_overlap(chromosome, start, end = NULL)
}
\arguments{
\item{chromosome}{Chromosome name (string). Example: "chr-13".}

\item{start}{Start position (integer).}

\item{end}{End position (integer). Default: NULL}
}
\value{
A data.frame containing TAD IDs, start, and end positions for TADs within the specified coordinates. If no data is available, a message is returned.
}
\description{
Function to retrieve human topologically associating domains (TADs) that completely overlap with a specific position or a specific range of genomic coordinates.
}
\examples{
\dontrun{
getTADs_by_overlap("chr-13", 34120000)
}

}
