% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf2prot.R
\name{mf2prot}
\alias{mf2prot}
\title{Molecular function to proteins (mf2prot)}
\usage{
mf2prot(molecular_function, taxon = NULL, sources = FALSE)
}
\arguments{
\item{molecular_function}{Molecular function label or ID (string). Example: "protein kinase activity".}

\item{taxon}{Taxon name (string) or taxon ID (string or integer), or NULL if not specified. Example: "Homo sapiens" or 9606. Default: NULL.}

\item{sources}{Include sources (boolean: TRUE or FALSE). Default: FALSE.}
}
\value{
A data.frame of proteins associated with the molecular function and their sources if sources is TRUE, otherwise a vector of unique protein names. If no data is available, a message is returned.
}
\description{
Function to obtain the proteins associated with a molecular function, optionally filtered by taxon and including sources.
}
\examples{
\dontrun{
mf2prot("protein kinase activity", "Homo sapiens", sources = TRUE)
}

}
