% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen2crm.R
\name{phen2crm}
\alias{phen2crm}
\title{Phenotype to CRMs (phen2crm)}
\usage{
phen2crm(phenotype)
}
\arguments{
\item{phenotype}{Phenotype label (string) or OMIM ID (string). Example: "schizophrenia" or "181500".}
}
\value{
A data.frame containing CRM names, OMIM IDs, databases, and articles associated with the phenotype, sorted by CRM name. If no data is available, a message is returned.
}
\description{
Function to obtain cis-regulatory modules (CRMs) associated with a phenotype, identified by a phenotype label or OMIM ID.
}
\examples{
\dontrun{
phen2crm("schizophrenia")
}

}
