% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen2gene.R
\name{phen2gene}
\alias{phen2gene}
\title{Phenotype to genes (phen2gene)}
\usage{
phen2gene(phenotype)
}
\arguments{
\item{phenotype}{Phenotype label (string) or OMIM ID (string). Example: "lung cancer".}
}
\value{
A data.frame containing gene names associated with the phenotype, sorted by gene name. If no data is available, a message is returned.
}
\description{
Function to retrieve genes associated with a given phenotype, identified by a phenotype label or OMIM ID.
}
\examples{
\dontrun{
phen2gene("lung cancer")
}

}
