% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot2bp.R
\name{prot2bp}
\alias{prot2bp}
\title{Protein to biological processes (prot2bp)}
\usage{
prot2bp(protein, taxon = "Homo sapiens", sources = F)
}
\arguments{
\item{protein}{Protein label or ID (string). Example: "TBX5_HUMAN" or "Q99593".}

\item{taxon}{Taxon name (string) or taxon ID (string or integer). Example: "Homo sapiens" or 9606. Default: "Homo sapiens".}

\item{sources}{Include sources (boolean: TRUE or FALSE). Default: FALSE.}
}
\value{
A data.frame containing biological process labels and IDs, and optionally relation labels, databases, and articles if sources is TRUE, sorted by process label. If no data is available, a message is returned.
}
\description{
Function to obtain biological processes associated with a specific protein, optionally filtered by taxon and including sources.
}
\examples{
\dontrun{
prot2bp("TBX5_HUMAN", taxon = "Homo sapiens", sources = FALSE)
}

}
