% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot2cc.R
\name{prot2cc}
\alias{prot2cc}
\title{Protein to cellular components (prot2cc)}
\usage{
prot2cc(protein)
}
\arguments{
\item{protein}{Protein label or ID (string). Example: "BRCA1_HUMAN" or "P38398".}
}
\value{
A data.frame containing cellular component IDs, labels, relation labels, databases, and articles, sorted by cellular component label. If no data is available, a message is returned.
}
\description{
Function to obtain cellular components associated with a specific protein, including their sources.
}
\examples{
\dontrun{
prot2cc("BRCA1_HUMAN")
}

}
