% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot2gene.R
\name{prot2gene}
\alias{prot2gene}
\title{Protein to gene (prot2gene)}
\usage{
prot2gene(protein)
}
\arguments{
\item{protein}{Protein label or ID (string). Example: "BRCA1_HUMAN" or "P38398".}
}
\value{
A data.frame containing gene IDs for the gene encoding the protein. If no data is available, a message is returned.
}
\description{
Function to retrieve the gene encoding a specific protein.
}
\examples{
\dontrun{
prot2gene("BRCA1_HUMAN")
}

}
