% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot2mf.R
\name{prot2mf}
\alias{prot2mf}
\title{Protein to molecular functions (prot2mf)}
\usage{
prot2mf(protein)
}
\arguments{
\item{protein}{Protein label or ID (string). Example: "BRCA1_HUMAN" or "P38398".}
}
\value{
A data.frame containing molecular function IDs, labels, relation labels, databases, and articles, sorted by molecular function label. If no data is available, a message is returned.
}
\description{
Function to obtain molecular functions associated with a specific protein, including their sources.
}
\examples{
\dontrun{
prot2mf("BRCA1_HUMAN")
}

}
