% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot2ortho.R
\name{prot2ortho}
\alias{prot2ortho}
\title{Protein to orthologous proteins (prot2ortho)}
\usage{
prot2ortho(protein)
}
\arguments{
\item{protein}{Protein label or ID (string). Example: "BRCA1_HUMAN" or "P38398".}
}
\value{
A data.frame containing orthologous protein labels, orthology relation labels, and taxon names, sorted by protein label. If no data is available, a message is returned.
}
\description{
Function to obtain orthologous proteins associated with a specific protein, including their taxa.
}
\examples{
\dontrun{
prot2ortho("BRCA1_HUMAN")
}

}
