% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot2prot.R
\name{prot2prot}
\alias{prot2prot}
\title{Protein to interacting proteins (prot2prot)}
\usage{
prot2prot(protein)
}
\arguments{
\item{protein}{Protein label or ID (string). Example: "BRCA1_HUMAN" or "P38398".}
}
\value{
A data.frame containing interacting protein labels, relation labels, databases, articles, evidence levels, and interaction details, sorted by protein label. If no data is available, a message is returned.
}
\description{
Function to obtain proteins that interact with a specific protein, including their sources and interaction details.
}
\examples{
\dontrun{
prot2prot("BRCA1_HUMAN")
}

}
