% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_chr.R
\name{translate_chr}
\alias{translate_chr}
\title{Translate chromosome name (translate_chr)}
\usage{
translate_chr(chromosome)
}
\arguments{
\item{chromosome}{Chromosome name (string). Example: "chr-1".}
}
\value{
A string containing the NCBI nuccore ID for the chromosome. If the chromosome name is invalid, a message is returned.
}
\description{
Function to translate a chromosome name to its corresponding NCBI nuccore ID.
}
\examples{
\dontrun{
translate_chr("chr-5")
}

}
