% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_data.R
\name{type_data}
\alias{type_data}
\title{Identify entity type (type_data)}
\usage{
type_data(instance)
}
\arguments{
\item{instance}{Entity label or ID (string). Example: "TP53" or "INSR_HUMAN".}
}
\value{
A character vector containing the type(s) of the entity. If no data is available, a message is returned.
}
\description{
Function to identify the type of a given entity.
}
\examples{
\dontrun{
type_data("TP53")
}

}
